requirements.txt
# Core LangGraph & LangChain Dependencies
langgraph>=0.2.45
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0

# Search & Web Tools
tavily-python>=0.5.0

# Data Validation & Models
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Environment Management
python-dotenv>=1.0.0

# UI Framework
streamlit>=1.40.0

# Utilities
requests>=2.32.0
aiohttp>=3.10.0

# Development Tools
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.0.0
flake8>=7.0.0

# Logging & Monitoring
loguru>=0.7.0
